+++ database/migrations/2026_04_23_190604_create_mesas_meseros_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('mesas_meseros', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('restaurante_id')->constrained('restaurantes')->onDelete('cascade');
-            $table->foreignId('propietario_id')->constrained('propietarios')->onDelete('cascade');
-            $table->foreignId('rol_id')->constrained('roles')->onDelete('cascade');
-            $table->integer('numero_mesa');
-            $table->timestamps();
-            
-            // Un mesero no debería tener la misma mesa duplicada en la tabla
-            $table->unique(['user_id', 'restaurante_id', 'numero_mesa']);
-        });
+        if (!Schema::hasTable('mesas_meseros')) {
+            Schema::create('mesas_meseros', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('restaurante_id')->constrained('restaurantes')->onDelete('cascade');
+                $table->foreignId('propietario_id')->constrained('propietarios')->onDelete('cascade');
+                $table->foreignId('rol_id')->constrained('roles')->onDelete('cascade');
+                $table->integer('numero_mesa');
+                $table->timestamps();
+                
+                // Un mesero no debería tener la misma mesa duplicada en la tabla
+                $table->unique(['user_id', 'restaurante_id', 'numero_mesa']);
+            });
+        }
     }
 
     /**
